+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * `bundled_project` are now loaded quietly (meaning, they are loaded, but not included into the projection)
 * `MongoQuerySettingsDict.pluck_from()`: now skips the `max_items` key because it does not make sense when inherited
 * Bugfix: in some cases, the `projection` property returned invalid results
+* Bugfix: when the `join` operation includes two relationships that are LEFT JOINed, the query is not broken anymore.
 
 ## 2.0.5 (2019-08-27)
 * `ensure_loaded` setting for projections